Updating DataSync specification.
Motivation:
The current DataSync specification contains information that is
outdated. The goal of this pull request is to document only what is
currently exists.

Modifications:
Updated the main docs/specs/aerogear-data-sync/ page so that it now only
contains a features section followed by links to the message format, the
java server api, and the java client api.

JIRA:
https://issues.jboss.org/browse/AGSYNC-31"